feature to allow copying mails to other folders, (configure via prefs -> ask for cofirmation before moving selected messages), using DIALOG BOXES by Michael Leigeber; suomi translations provided by mkk
git-svn-id: http://svn.egroupware.org/egroupware/trunk@29998 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ felamimail/inc/class.bofelamimail.inc.php
@@ -2557,7 +2557,7 @@
 			}
 		}
 
-		function moveMessages($_foldername, $_messageUID)
+		function moveMessages($_foldername, $_messageUID, $deleteAfterMove=true)
 		{
 			$msglist = '';
 
@@ -2566,16 +2566,19 @@
 			if ( PEAR::isError($this->icServer->copyMessages($_foldername, $_messageUID, $this->sessionData['mailbox'], true)) ) {
 				return false;
 			}
-			// mark messages as deleted
-			if ( PEAR::isError($this->icServer->deleteMessages($_messageUID, true))) {
-				return false;
-			}
+			if ($deleteAfterMove === true)
+			{
+				if ( PEAR::isError($this->icServer->deleteMessages($_messageUID, true))) 
+				{
+					return false;
+				}
 
-			if($deleteOptions != "mark_as_deleted") {
-				// delete the messages finaly
-				$this->icServer->expunge();
-			}
-
+				if($deleteOptions != "mark_as_deleted") 
+				{
+					// delete the messages finaly
+					$this->icServer->expunge();
+				}
+ 			}
 			return true;
 		}
